dom/security/test/csp/test_worker_redirect.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Bug 949706 - CSP: Correct handling of web workers importing scripts that get redirected</title>
  <!-- Including SimpleTest.js so we can use waitForExplicitFinish !-->
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
  <div id="content" style="visibility: hidden">
    <iframe style="width:100%;" id="testframe"></iframe>
  </div>

<script class="testbody" type="text/javascript">

/* Description of the test:
 * We load a page that loads a script which then instantiates a web worker,
 * where that web worker then imports a script which gets redirected.
 * We verify that the CSP applies correctly after the imported script of
 * the worker gets redirected. More specifically, the test works as follows:
 *
 * test_worker_redirect.html
 * -> loads file_worker_redirect.html file into iframe
 * -> loads worker file_worker_redirect.sjs?stage_0_script_loads_worker
 * -> creates script file_worker_redirect.sjs?stage_1_worker_import_scripts
 * -> redirects script file_worker_redirect.sjs?stage_2_redirect_imported_script
 * -> loads target script file_worker_redirect.sjs?stage_3_target_script
 *
 * Please note that we have to use 'unsafe-eval' in the policy
 * so that workers are actually permitted by the CSP.
 *
 * The main test is loaded using:
 *   http://mochi.test:8888
 * where the imported script gets redirected to:
 *   http://test1.example.com
 */

var tests = [
  {
    policy: "default-src 'self'; script-src 'self' 'unsafe-eval'; child-src 'self' http://test1.example.com;",
    expected: "allowed"
  },
  {
    policy: "default-src 'self'; script-src 'self' 'unsafe-eval'; child-src 'self';",
    expected: "blocked",
  },
];

var counter = 0;
var curTest;

function checkResult(aResult) {
  is(aResult, curTest.expected, "Should be (" + curTest.expected + ") in Test " + counter + "!");
  loadNextTest();
}

function loadNextTest() {
  if (counter == tests.length) {
    SimpleTest.finish();
    return;
  }
  curTest = tests[counter++];
  var src = "file_testserver.sjs";
  // append the file that should be served
  src += "?file=" + escape("tests/dom/security/test/csp/file_worker_redirect.html");
  // append the CSP that should be used to serve the file
  src += "&csp=" + escape(curTest.policy);
  document.getElementById("testframe").src = src;
}

SimpleTest.waitForExplicitFinish();
loadNextTest();

</script>
</body>
</html>
